requirements.txt
# Core dependencies for Unsupervised Elicitation
# vLLM provides: aiohttp, anthropic, fastapi, numpy, openai, pydantic, requests,
#                scipy, tiktoken, tokenizers, torch, tqdm, transformers

# Data processing & analysis
adjustText
alembic
altair
cattrs
datasets  # HuggingFace datasets
fire
hydra-core
matplotlib
pandas
polars
scikit-learn
seaborn
sqlalchemy==2.0.18
tabulate
trueskill

# LLM APIs (not covered by vLLM)
mistralai
replicate

# Utilities
pebble
pre-commit
tenacity
termcolor
typer[all]
wandb

# vLLM for batched inference (no version pin - use latest 0.13.x)
vllm>=0.13.0,<0.14.0
